# swagger.yaml
openapi: 3.0.3
info:
  title: AXZY CHECK API
  description:
    Documentación técnica completa de la API de AXZY CHECK. Cubre gestión
    operativa, administrativa y analítica.
  version: 1.3.2
servers:
  - url: http://localhost:4444/api/v1
    description: Local server
...
paths:
...
  /incidents:
    get:
      tags:
        - incidents
      summary: Listar incidencias
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResult"
    post:
      tags:
        - incidents
      summary: Crear reporte de incidencia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentCreate"
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResult"
        "400":
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResultError"
...
  /incidents/{id}:
    delete:
      tags:
        - incidents
      summary: Eliminar incidencia
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResultListRound"
  /maintenance:
    get:
      tags:
        - maintenance
      summary: Listar solicitudes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResultListMaintenance"
    post:
      tags:
        - maintenance
      summary: Crear solicitud
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceCreate"
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResultDatatableRound"
        "400":
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResultError"
...
  /maintenance/{id}:
    delete:
      tags:
        - maintenance
      summary: Eliminar mantenimiento
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResultDatatableIncident"
        "400":
          description: Petición inválida o restricción de llave foránea
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResultError"
        "404":
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResultError"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TResultError"
...
components:
  schemas:
...
    IncidentCreate:
      type: object
      required:
        - title
        - locationId
      properties:
        title:
          type: string
        description:
          type: string
        locationId:
          type: string
          format: uuid
        media:
          type: array
          items:
            type: string
    MaintenanceCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        description:
          type: string
        media:
          type: array
          items:
            type: string
...
    Round:
      type: object
      properties:
        id: &id127
          type: string
          format: uuid
        guardId: &id128
          type: string
          format: uuid
        startTime: &id129
          type: string
          format: date-time
        endTime: &id130
          type: string
          format: date-time
          nullable: true
        status:
          $ref: "#/components/schemas/RoundStatus"
        recurringConfigurationId: &id131
          type: string
          format: uuid
          nullable: true
        createdAt: &id132
          type: string
          format: date-time
        updatedAt: &id133
          type: string
          format: date-time
        deletedAt: &id134
          type: string
          format: date-time
          nullable: true
        guard:
          $ref: "#/components/schemas/UserBasic"
        recurringConfiguration:
          $ref: "#/components/schemas/RecurringConfigurationBasic"
...
    Incident:
      type: object
      properties:
        id: &id113
          type: string
          format: uuid
        guardId: &id114
          type: string
          format: uuid
        title: &id115
          type: string
        categoryId: &id116
          type: string
          format: uuid
          nullable: true
        typeId: &id117
          type: string
          format: uuid
          nullable: true
        description: &id118
          type: string
          nullable: true
        media: &id119
          type: object
          nullable: true
        latitude: &id120
          type: number
          nullable: true
        longitude: &id121
          type: number
          nullable: true
        createdAt: &id122
          type: string
          format: date-time
        updatedAt: &id123
          type: string
          format: date-time
        deletedAt: &id124
          type: string
          format: date-time
          nullable: true
        resolvedAt: &id125
          type: string
          format: date-time
          nullable: true
        resolvedById: &id126
          type: string
          format: uuid
          nullable: true
        status:
          $ref: "#/components/schemas/IncidentStatus"
        category:
          $ref: "#/components/schemas/IncidentCategoryBasic"
        type:
          $ref: "#/components/schemas/IncidentTypeBasic"
        guard:
          $ref: "#/components/schemas/UserBasic"
        resolvedBy:
          $ref: "#/components/schemas/UserBasic"
    Maintenance:
      type: object
      properties:
        id: &id159
          type: string
          format: uuid
        guardId: &id160
          type: string
          format: uuid
        title: &id161
          type: string
        categoryId: &id162
          type: string
          format: uuid
          nullable: true
        typeId: &id163
          type: string
          format: uuid
          nullable: true
        category: &id164
          type: string
          nullable: true
        description: &id165
          type: string
          nullable: true
        media: &id166
          type: object
          nullable: true
        latitude: &id167
          type: number
          nullable: true
        longitude: &id168
          type: number
          nullable: true
        status:
          $ref: "#/components/schemas/MaintenanceStatus"
        resolvedAt: &id169
          type: string
          format: date-time
          nullable: true
        resolvedById: &id170
          type: string
          format: uuid
          nullable: true
        createdAt: &id171
          type: string
          format: date-time
        updatedAt: &id172
          type: string
          format: date-time
        deletedAt: &id173
          type: string
          format: date-time
          nullable: true
        guard:
          $ref: "#/components/schemas/UserBasic"
        resolvedBy:
          $ref: "#/components/schemas/UserBasic"
        categoryRel:
          $ref: "#/components/schemas/IncidentCategoryBasic"
        type:
          $ref: "#/components/schemas/IncidentTypeBasic"
...
    TResult:
      type: object
      description: Estructura de respuesta estándar AXZY
      properties:
        success:
          type: boolean
          description: Indica si la operación fue exitosa
          example: true
        data:
          type: object
          description: Datos de la respuesta
          example:
            id: uuid
            name: Ejemplo
        messages:
          type: array
          description: Lista de mensajes informativos o de éxito
          items:
            type: string
          example:
            - Operación realizada con éxito
...
    TResultError:
      type: object
      description: Estructura de respuesta de error estándar
      properties:
        success:
          type: boolean
          example: false
        data:
          type: object
          nullable: true
          example: null
        messages:
          type: array
          description: Detalles del error o validación
          items:
            type: string
          example:
            - Credenciales inválidas
            - El campo nombre es obligatorio
        stack:
          type: string
          description: Traza del error (solo en desarrollo)
          example: "Error: Stack trace sample..."
...
    TResultDatatableRound:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            rows:
              type: array
              items:
                $ref: "#/components/schemas/Round"
            total:
              type: integer
              example: 100
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
        messages:
          type: array
          items:
            type: string
    TResultListRound:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Round"
        messages:
          type: array
          items:
            type: string
    TResultDatatableIncident:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            rows:
              type: array
              items:
                $ref: "#/components/schemas/Incident"
            total:
              type: integer
              example: 100
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
        messages:
          type: array
          items:
            type: string
...
    TResultListMaintenance:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Maintenance"
        messages:
          type: array
          items:
            type: string
...
    IncidentStatus:
      type: string
      enum:
        - PENDING
        - ATTENDED
    RoundStatus:
      type: string
      enum:
        - IN_PROGRESS
        - COMPLETED
    MaintenanceStatus:
      type: string
      enum:
        - PENDING
        - ATTENDED
...
    UserBasic:
      type: object
      properties:
        id: *id014
        name: *id015
        lastName: *id016
        username: *id017
        active: *id018
        softDelete: *id019
        shiftStart: *id020
        shiftEnd: *id021
        createdAt: *id022
        updatedAt: *id023
        deletedAt: *id024
        isLoggedIn: *id025
        scheduleId: *id026
        roleId: *id027
...
    IncidentCategoryBasic:
      type: object
      properties:
        id: *id097
        name: *id098
        value: *id099
        type: *id100
        color: *id101
        icon: *id102
        createdAt: *id103
        updatedAt: *id104
        deletedAt: *id105
...
    IncidentTypeBasic:
      type: object
      properties:
        id: *id106
        categoryId: *id107
        name: *id108
        value: *id109
        createdAt: *id110
        updatedAt: *id111
        deletedAt: *id112
...
    RecurringConfigurationBasic:
      type: object
      properties:
        id: *id138
        title: *id139
        active: *id140
        softDelete: *id141
        createdAt: *id142
        updatedAt: *id143
        deletedAt: *id144
...